% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{absmax}
\alias{absmax}
\title{Find Maximum Magnitude Element of Vector}
\usage{
absmax(x)
}
\arguments{
\item{x}{A vector}
}
\value{
the signed maximum
}
\description{
Return the (signed) value of the maximum (in magnitude) element of a vector
}
\details{
An example is especially illustrative. absmax(c(1,-3,2)) will
yield -3. absmax(c(1,-2,3)) will yield 3.
}
\author{
Daniel Kessler
}
