% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{bt_algo_asympt_inner}
\alias{bt_algo_asympt_inner}
\title{Batchtools Algorithm: Asymptotic CIs (Inrep Function)}
\usage{
bt_algo_asympt_inner(inrep, level, align)
}
\arguments{
\item{inrep}{A given internal replication of the problem}

\item{level}{Value in (0,1), e.g., 0.95 is a 95\% CI}

\item{align}{A function to use to align. Can make use of inrep$fm_hat}
}
\value{
A list with two objects: xcoef_ci and ycoef_ci
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Daniel Kessler
}
