% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{bt_algo_boot}
\alias{bt_algo_boot}
\title{Batchtools Function for Bootstrapped Confidence Intervals}
\usage{
bt_algo_boot(job, data, instance, ...)
}
\arguments{
\item{job}{required by batchtools}

\item{data}{required by batchtools}

\item{instance}{required by batchtools}

\item{...}{Arguments passed on to ci_func}
}
\value{
CI object
}
\description{
Batchtools Function for Bootstrapped Confidence Intervals
}
\author{
Dan Kessler
}
