% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{bt_algo_bootabs_fun}
\alias{bt_algo_bootabs_fun}
\title{BT algo fun for Absolute Value Bootstrap}
\usage{
bt_algo_bootabs_fun(
  job = NULL,
  data = NULL,
  instance = NULL,
  level = 0.95,
  nboots = 1000,
  parametric = FALSE
)
}
\arguments{
\item{job}{}

\item{data}{}

\item{instance}{}
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Daniel Kessler
}
