% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{bt_algo_bootabs_inner}
\alias{bt_algo_bootabs_inner}
\title{}
\usage{
bt_algo_bootabs_inner(inrep)
}
\value{

}
\description{
Given a single inrep, do your thing
}
\details{
.. content for \details{} ..
}
\author{
Daniel Kessler
}
