% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{bt_algo_inrep}
\alias{bt_algo_inrep}
\title{Generic inrep function for batchtools algorithms}
\usage{
bt_algo_inrep(inrep, ci_func, ...)
}
\arguments{
\item{inrep}{The inrep object}

\item{ci_func}{A function that follows the cca_ci_* API}

\item{...}{Arguments passed on to ci_func}
}
\value{
CIs as returned by ci_func
}
\description{
Generic inrep function for batchtools algorithms
}
\author{
Dan Kessler
}
