% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cancor.cov}
\alias{cancor.cov}
\title{Canonical Correlation Analysis for Covariance Matrix}
\usage{
cancor.cov(Sigma, px)
}
\arguments{
\item{Sigma}{A PSD matrix given the covariance (sample or
population)}

\item{px}{The number of variables in the x dataset}
}
\value{
A list containing the following components:

cor: correlations

xcoef: estimated coefficients for the x variables

ycoef: estimated coefficients for the y variables
}
\description{
Compute the Canonical Correlation Analysis solution based on a
covariance matrix
}
\details{
The Canonical Correlation Analysis problem can be solved in a
variety of ways. The approach taken by the cancor function in R
performs QR decomposition on the data matrices and proceeds from
there. However, it is also possible to find the canonical
correlations and associated weights using only the covariance
matrix. This is especially useful if one wishes to solve the
"population" version of CCA for a distribution with population
covariance Sigma. This algorithm results in canonical variates
that have unit empirical variance.

Note: As a convenience to the user, this function will also fix
the signs of the resulting coefficients as described in
cancor.signfix.
}
\author{
Daniel Kessler
}
