% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cancor_scaled}
\alias{cancor_scaled}
\title{Canonical Correlations}
\usage{
cancor_scaled(x, y, xcenter = TRUE, ycenter = TRUE)
}
\value{

}
\description{
Compute the canonical correlations between two data matrices such that
canonical variates have unit variance.
}
\details{
This is a simple wrapper around stats::cancor. stats::cancor uses the Golub
and Van Loan algorithm, which yields canonical variates that have unit norm
rather than unit variance.
As a result, the canonical variates will have (empirical) variance
of 1/(N-1), where N is the number of observations. This is
undesirable if we wish to perform inference on the weights in
xcoef or ycoef, because their scale will vary with N.

This function simply calls stats::cancor and then multiplies both xcoef and
ycoef by sqrt(N-1) so that the resulting canonical variates will indeed have
unit variance.

See the documentation for stats::cancor for more details on standard usage.
}
\author{
Daniel Kessler
}
