% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cancor_scalefix}
\alias{cancor_scalefix}
\title{Fix Cancor Scaling}
\usage{
cancor_scalefix(fm, N)
}
\arguments{
\item{fm}{A fitted object returned by cancor}

\item{N}{The number of observations}
}
\value{
A modified version of fm
}
\description{
Adjust canonical correlation analysis results so that canonical
variates have unit variance
}
\details{
Cancor uses the Golub and Van Loan algorithm, which yields
canonical variates that have unit norm rather than unit variance.
As a result, the canonical variates will have (empirical) variance
of 1/(N-1), where N is the number of observations. This is
undesirable if we wish to perform inference on the weights in
xcoef or ycoef, because their scale will vary with N.

This function simply multiplies both xcoef and ycoef by sqrt(N-1)
so that the resulting canonical variates will indeed have unit variance.
}
\author{
Daniel Kessler
}
