% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{cancor_signfix}
\alias{cancor_signfix}
\title{Fix Cancor Signs}
\usage{
cancor_signfix(fm)
}
\arguments{
\item{fm}{A fitted object returned by cancor}
}
\value{
Same object as returned by cancor after sign-flipping per
the identifiability condition discussed in Details.
}
\description{
Correct sign ambiguity in canonical correlation analysis
}
\details{
Canonical Correlation Analysis suffers from sign ambiguity in the
estimated coefficients. This is because Corr(U, V) = Corr(-U, -V).
This function adopts the convention that each canonical
coefficient pair (xcoef, ycoef) should satisfy the condition that
its maximal element (in absolute value) has positive sign. You may
wish to standardize the columns of X and Y prior to initially
fitting CCA so that differences in variable scaling do not
dominate the identity of the maximal element.

Note: This only cares about and tries to fix sign ambiguities in
the first K=min(px,py) components.
}
\author{
Daniel Kessler
}
