% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cancor_signfix_diag}
\alias{cancor_signfix_diag}
\title{Fix cancor signs based on the diagonal}
\usage{
cancor_signfix_diag(fm)
}
\arguments{
\item{fm}{A fitted object returned by cancor (or similar)}
}
\value{
Object like fm but with possible sign flips
}
\description{
Correct sign ambiguity in canonical correlation analysis by requiring that
the diagonal of xcoef be non-negative (this is the approach suggested by
Anderson 1999). This will also truncate both xcoef and ycoef to have the
same number of columns as the length of cor
}
\author{
Dan Kessler
}
