% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_align_nil}
\alias{cca_align_nil}
\title{Non-functional CCA alignment}
\usage{
cca_align_nil(fm, ref)
}
\description{
Align CCA solutions by doing nothing.
}
\details{
.. content for \details{} ..
}
\author{
Dan Kessler
}
