% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_align_posdiag}
\alias{cca_align_posdiag}
\title{Fix cancor signs based on the diagonal}
\usage{
cca_align_posdiag(fm, ref)
}
\arguments{
\item{fm}{A fitted object returned by cancor (or similar)}

\item{ref}{Not used}
}
\value{
Object like fm but with possible sign flips
}
\description{
Correct sign ambiguity in canonical correlation analysis by requiring that
the diagonal of xcoef be non-negative (this is the approach suggested by
Anderson 1999). This will also truncate both xcoef and ycoef to have the
same number of columns as the length of cor
}
\author{
Dan Kessler
}
