% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_boot}
\alias{cca_ci_boot}
\title{Bootstrap-based confidence intervals for CCA}
\usage{
cca_ci_boot(
  x,
  y,
  level = 0.9,
  align = cca_align_nil,
  ref,
  nboots = 100,
  ncpus = 1,
  boot_type = c("norm", "basic", "perc", "bca"),
  return_boot = FALSE
)
}
\arguments{
\item{x}{Data matrix of size n by p}

\item{y}{Data matrix of size n by q}

\item{level}{Level for confidence intervals, should be in (0, 1)}

\item{align}{Function for alignment}

\item{ref}{Passed through to align function}

\item{nboots}{How many bootstrap replicates}

\item{ncpus}{How many cpu's to use to parallelize bootstrap. If > 1, then
"multicore" will be used}

\item{boot_type}{Which bootstrap methods to use; see type argument of
boot::boot.ci}

\item{return_boot}{Whether to return the (potentially very large) boot_out
object created by the bootstrap function. Defaults to FALSE}
}
\value{
List of several types of CIs
}
\description{
Use the boot package to get bootstrapped confidence intervals for CCA
}
\details{
Details go here
}
\author{
Dan Kessler
}
