% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{cca_ci_bootstrap}
\alias{cca_ci_bootstrap}
\title{Bootstrap-based confidence intervals for CCA directions}
\usage{
cca_ci_bootstrap(
  x,
  y,
  level = 0.05,
  nboots = 1000,
  parametric = FALSE,
  progress = 0
)
}
\arguments{
\item{x}{Data matrix of size N by p}

\item{y}{Data matrix of size N by q}

\item{level}{Level for confidence intervals, should be in (0, 1)}

\item{nboots}{Number of bootstrap sample to draw}

\item{parametric}{If FALSE (default), do bootstrap by sampling with
replacement. If TRUE, perform parametric bootstrap, i.e., draw data from
multivariate normal distribution following sample covariance}

\item{progress}{If 0 (default), don't report progress. If set to a positive
integer k, report bootstrap progress every k'th run.}
}
\value{

}
\description{
Bootstrap data and
}
\details{
.. content for \details{} ..
}
\author{
Daniel Kessler
}
