% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_bootstrap_abs}
\alias{cca_ci_bootstrap_abs}
\title{Bootstrap-based confidence intervals for CCA directions}
\usage{
cca_ci_bootstrap_abs(
  x,
  y,
  level = 0.95,
  align,
  ref,
  nboots = 1000,
  parametric = FALSE,
  progress = 0
)
}
\arguments{
\item{x}{Data matrix of size N by p}

\item{y}{Data matrix of size N by q}

\item{level}{Level for confidence intervals, should be in (0, 1)}

\item{nboots}{Number of bootstrap sample to draw}

\item{parametric}{If FALSE (default), do bootstrap by sampling with
replacement. If TRUE, perform parametric bootstrap, i.e., draw data from
multivariate normal distribution following sample covariance}

\item{progress}{If 0 (default), don't report progress. If set to a positive
integer k, report bootstrap progress every k'th run.}
}
\value{
List with two objects: xcoef_ci and ycoef_ci.
}
\description{
Bootstrap absolute quantiles to generate CCA confidence intervals
}
\author{
Daniel Kessler
}
