% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_coverage1}
\alias{cca_ci_coverage1}
\title{Compute CI coverage for a vector}
\usage{
cca_ci_coverage1(true, cis)
}
\arguments{
\item{true}{A vector of length p}

\item{cis}{A matrix of size p x 2; first column is lower bound of CI, second
column is upper bound of CI}
}
\value{
A scalar in \link{0, 1} with the proportion of covered intervals
}
\description{
Compute CI coverage for a vector
}
\author{
Dan Kessler
}
