% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_coverage_pangloss}
\alias{cca_ci_coverage_pangloss}
\title{Compute the best possible coverage of CCA CIs (allowing both sign
flips and reassignments)}
\usage{
cca_ci_coverage_pangloss(fm_true, cis, xyweight = NULL)
}
\arguments{
\item{xyweight}{A number in \link{0, 1} indicating how much to weigh coverage of
x's directions vs y's. If NULL, they are weighted by p and q. If 0, this
only cares about matching x, if 1, only cares about matching y.}
}
\value{
Mean coverage
}
\description{
Compute the best possible coverage of CCA CIs (allowing both sign
flips and reassignments)
}
\author{
Dan Kessler
}
