% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_coverage_possibilities}
\alias{cca_ci_coverage_possibilities}
\title{Compute All Possible Coverages of CCA CIs}
\usage{
cca_ci_coverage_possibilities(fm_true, cis)
}
\arguments{
\item{fm_true}{The result of cancor_cov}

\item{cis}{A list with xcoef and ycoef, which are p x k x 2 and q x k x 2,
respectively}
}
\value{
A list containing four k x k matrices of coverage by components
under various assignment schemes. Rows index components of fm_true,
columns index components of cis. xcov_pos holds coverage for x without
sign flipping, xcov_neg holds coverage for x after a -1 sign flip,
ycov_pos and ycov_neg are analaogous.
}
\description{
Compute All Possible Coverages of CCA CIs
}
\author{
Dan Kessler
}
