% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_lengths}
\alias{cca_ci_lengths}
\title{Compute the lengths of CCA CIs}
\usage{
cca_ci_lengths(cis)
}
\arguments{
\item{cis}{A list with xcoef_ci and ycoef_ci, which are each arrays with
dimensions p x K x 2 and q x K x 2, respectively}
}
\value{
A list with xcoef_lengths and ycoef_lengths
}
\description{
Compute the lengths of CCA CIs
}
\author{
Dan Kessler
}
