% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{cca_ci_regression}
\alias{cca_ci_regression}
\title{Regression-Based Confidence Intervals for CCA Directions}
\usage{
cca_ci_regression(x, y, level = 0.95, train_ratio = 0.5)
}
\arguments{
\item{x}{Data matrix of size n by p}

\item{y}{Data matrix of size n by q}

\item{level}{Level for confidence intervals, should be in (0, 1)}

\item{train_ratio}{What proportion of the data should be used for estimating
CCA directions}
}
\value{
List with two objects: xcoef_ci and ycoef_ci.
}
\description{
Obtain confidence intervals for canonical correlation analysis directions by
taking a regression-approach
}
\details{
The idea of the regression approach is to recast the CCA problem as a
regression problem. First, select a proportion of the data equal to
\code{train_ratio} of the data (rounding down as necessary). Then, estimate
the usual CCA parameters on this partition: these include \code{cor} (the
canonical correlation coefficients), \code{xcoef}, (the canonical directions
associated with x), and \code{ycoef} (the canonical directions associated
with y).

Use the canonical directions estimated above to construct new canonical
variables in the held-out partition (by taking the inner product between
each observation and a given canonical direction). Then, fit many regression
models where we predict one canonical variable using the other dataset. For
each, apply a transformation such that the predicted values have unit
variance (to satisfy the CCA constraint). Finally, use conventional
regression-based inference to provide confidence intervals for the
coefficients.
}
\author{
Dan Kessler
}
