% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{coverage_experiment}
\alias{coverage_experiment}
\title{Conduct asymptotic coverage experiment}
\usage{
coverage_experiment(
  outreps = 1L,
  inreps = 1L,
  p = 2,
  q = 2,
  n = 1000L,
  nboots = 1000L
)
}
\arguments{
\item{outreps}{Each "outer replication" draws a new value of sigma. This can
either be an integer (e.g., 10L), in which case a random covariance matrix
will be drawn for each out replication, or instead it can be a list whose
length determines the number of replications, and whose values are
covariance matrices to use.}

\item{inreps}{Each "inner replication" is a repetition with new data for the
same value of sigma}

\item{p}{The dimension of X}

\item{q}{The dimension of Y}

\item{n}{How many datapoints to draw in each sample}

\item{nboots}{How many bootstrap samples to draw}

\item{sigma}{Optional. A list of length outreps containing covariance
matrices of size p + q. This matrices will be used for each of the outreps
replications. If not provided, a covariance matrix will be generated for
you for each of the outreps.}
}
\value{
A list of results, containing: (1) truth, a 4D array of true
parameter values, (2) cover, a 5D array of logicals indicating which CIs
cover the truth, (3) length, a 5D array of CI lengths, (4) cis, a 6D array
of the confidence intervals, and (5) sigma, a list of length outreps the
holds the generative Sigma. For the arrays, the dimensions (as applicable)
index (i) coordinates (p+q), (ii) components (K), (iii) inreps, (iv)
outreps, (v) method, (vi) lower/upper confidence bounds.
}
\description{
Conduct asymptotic coverage experiment
}
\author{
Dan Kessler
}
