% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{gen_data}
\alias{gen_data}
\title{Draw CCA data according to covariance Sigma}
\usage{
gen_data(Sigma, p, q, n)
}
\arguments{
\item{Sigma}{Square, positive definite covariance matrix with p + q
rows/cols}

\item{p}{The dimension of the random variable X}

\item{q}{The dimension of the random variable Y}

\item{n}{The number of observations to draw}
}
\value{
A list with two fields: x and y. They are matrices of size n by p
and n by q, respectively.
}
\description{
Draw data for use in CCA estimation according to a user-specified covariance
matrix.
}
\author{
Dan Kessler
}
