% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{gen_sigma2}
\alias{gen_sigma2}
\title{Generate covariance matrix for CCA}
\usage{
gen_sigma2(p, q, cov_type, rho1, rho2, type = "sparse")
}
\arguments{
\item{p}{Dimension of X random variable}

\item{q}{Dimension of Y random variable}

\item{cov_type}{Either "id" or "sPrec"}

\item{type}{Either "sparse", "constant", or "random"}

\item{rho_1}{The second-largest canonical correlation}
}
\value{
A square, positive definite matrix with p+q rows/cols
}
\description{
Generate covariance matrix for CCA
}
\author{
Dan Kessler
}
