% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{lm_scalefix}
\alias{lm_scalefix}
\title{Refit a linear model such that predicted values have unit variance}
\usage{
lm_scalefix(fm)
}
\arguments{
\item{fm}{An object fitted with lm}
}
\value{
An object of type lm refit such that predicted values have unit
variance
}
\description{
Refit a linear model such that predicted values have unit variance
}
\author{
Dan Kessler
}
