% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{bt_problem_std_fun}
\alias{bt_problem_std_fun}
\title{Batchtools convenience function for "problem"}
\usage{
bt_problem_std_fun(
  job = NULL,
  data = NULL,
  sigma = NULL,
  p = NULL,
  q = NULL,
  n = NULL,
  inreps = 1L,
  ...
)
}
\arguments{
\item{job}{Constructed by batchtools}

\item{data}{Optionally, pass a list with named elements \code{x} and \code{y}, where
each is an n by p and n by q matrix, respectively}

\item{sigma}{Optional: a square covariance matrix of dimension (p + q)}

\item{p}{The dimension of the random variable X}

\item{q}{The dimension of the random variable Y}

\item{n}{The number of observations}

\item{inreps}{How many replicates to draw}

\item{...}{Additional arguments passed to gen_sigma}
}
\value{
An instance
}
\description{
This is a convenience function for use with batchtools
}
\details{
The details go here
}
\author{
Daniel Kessler
}
