% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{bt_problem_std_fun}
\alias{bt_problem_std_fun}
\title{Batchtools convenience function for "problem"}
\usage{
bt_problem_std_fun(
  job = NULL,
  data = NULL,
  data_is_sigma = FALSE,
  pq = NULL,
  n = NULL,
  cov_type = c("id", "sPrec")[1],
  rho1 = 0.9,
  rho2 = 0,
  fiddle.var = NULL,
  fiddle.comp = NULL,
  fiddle.val = NULL,
  inreps = 1L
)
}
\arguments{
\item{job}{Constructed by batchtools}

\item{data}{Optionally, pass a list with named elements \code{x} and \code{y}, where
each is an n by p and n by q matrix, respectively. Or, if \code{sigma} is true,
then pass a single square covariance matrix of dimension (p + q)}

\item{data_is_sigma}{If true, treat "data" as sigma: a square covariance
matrix of dimension (p + q)}

\item{pq}{A string like "p,q" with dimensions of the random variables X, Y}

\item{n}{The number of observations}

\item{cov_type}{Specifies covariance structure for X and Y. If "id", then
both have identity covariance matrices. If "sPrec", then both will
covariance corresponding to sparse precision matrices (see details).}

\item{rho1}{Largest canonical correlation}

\item{rho2}{Second largest canonical correlation}

\item{fiddle.var}{Coefficients of which variable you want to modify}

\item{fiddle.comp}{Component whose last coordinate you want to modify}

\item{fiddle.val}{String, one of "0", "mean", or "max"}

\item{inreps}{}
}
\value{
An instance
}
\description{
This is a convenience function for use with batchtools
}
\details{
TODO: Write details for sPrec
}
\author{
Daniel Kessler
}
