% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cancor_scaled}
\alias{cancor_scaled}
\title{Canonical Correlations with Unit Variance}
\usage{
cancor_scaled(
  x,
  y,
  xcenter = TRUE,
  ycenter = TRUE,
  align = cca_align_posdiag,
  ref
)
}
\arguments{
\item{x}{one dataset}

\item{y}{the other dataset}

\item{xcenter}{whether to center x}

\item{ycenter}{whether to center y}

\item{align}{A function for aligning the solution}

\item{ref}{Passed through to align}
}
\description{
Compute the CCA with unit variance constraint
}
\details{
This is a simple wrapper around stats::cancor. stats::cancor uses the Golub
and Van Loan algorithm, which yields canonical variates that have unit norm
rather than unit variance.
As a result, the canonical variates will have (empirical) variance
of 1/(N-1), where N is the number of observations. This is
undesirable if we wish to perform inference on the weights in
xcoef or ycoef, because their scale will vary with N.

This function calls stats::cancor and then multiplies both xcoef and ycoef
by sqrt(N-1) so that the resulting canonical variates will indeed have unit
variance.

In addition, this function adds two new fields to the result, which hold the
columnwise standard deviations of x and y, respectively. These fields may be
useful for downstream tasks (e.g., alignment).

See the documentation for stats::cancor for more details on standard usage.
}
\author{
Daniel Kessler
}
