% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_asymptotic}
\alias{cca_ci_asymptotic}
\title{Asymptotic confidence intervals for CCA directions}
\usage{
cca_ci_asymptotic(x, y, level = 0.95, align = cancor_signfix_diag)
}
\arguments{
\item{x}{Data matrix of size n by p}

\item{y}{Data matrix of size n by q}

\item{level}{Level for confidence intervals, should be in (0, 1)}

\item{align}{A function to perform post-processing on the estimated
coefficients to render the solution well-identified. By default, this uses
cancor_signfix_diag, which ensures that the diagonal of xcoef is
non-negative.}
}
\value{
List with two objects: xcoef_ci and ycoef_ci.
}
\description{
Obtain confidence intervals for the "directions" of a canonical correlation
analysis using asymptotic results from Anderson 1999.
}
\author{
Dan Kessler
}
