% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_boot}
\alias{cca_ci_boot}
\title{Bootstrap-based confidence intervals for CCA}
\usage{
cca_ci_boot(x, y, level = 0.9, align = cca_align_posdiag, ref, nboots = 100)
}
\arguments{
\item{x}{Data matrix of size n by p}

\item{y}{Data matrix of size n by q}

\item{level}{Level for confidence intervals, should be in (0, 1)}

\item{align}{Function for alignment}

\item{ref}{Passed through to align function}

\item{nboots}{How many bootstrap replicates}
}
\value{
List of several types of CIs
}
\description{
Use the boot package to get bootstrapped confidence intervals for CCA
}
\details{
Details go here
}
\author{
Dan Kessler
}
