% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cca_ci_coverage_signflip}
\alias{cca_ci_coverage_signflip}
\title{Compute the coverage of CCA CIs allowing sign flips}
\usage{
cca_ci_coverage_signflip(fm_true, cis, xyweight = NULL)
}
\arguments{
\item{fm_true}{}

\item{cis}{}

\item{xyweight}{}
}
\value{
Mean coverage
}
\description{
Compute the coverage of CCA CIs allowing sign flips
}
\author{
Dan Kessler
}
