% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{cos_sim}
\alias{cos_sim}
\title{Compute cosine similarity between columns of two matrices}
\usage{
cos_sim(x, y)
}
\arguments{
\item{x}{Matrix of size n by p}

\item{y}{Matrix of size n by q}
}
\value{
Matrix of size p by q of cosine similarities. The (i, j) entry has
the similarity between x_i and y_j.
}
\description{
Compute cosine similarity between columns of two matrices
}
\author{
Daniel Kessler

Note that the cosine similarity is not well-defined when there exists a
column of all 0's in either x or y.
}
