% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{coverage_experiment}
\alias{coverage_experiment}
\title{Conduct asymptotic coverage experiment}
\usage{
coverage_experiment(outreps, inreps, p, q, n)
}
\arguments{
\item{outreps}{Each "outer replication" draws a new value of sigma}

\item{inreps}{Each "inner replication" is a repetition with new data for the
same value of sigma}

\item{p}{The dimension of X}

\item{q}{The dimension of Y}

\item{n}{How many datapoints to draw in each sample}
}
\value{
A big array with all of the results. The array has the following
dimensions: outreps by inreps by (p + q) by K by 3. The last dimension
holds the lower CI estimates, upper CI estimates, and true values,
respectively.
}
\description{
Conduct asymptotic coverage experiment
}
\author{
Dan Kessler
}
